[gentest] Refactor gentest to use WebDriver & quote crate.
gentest would not build with a current openssl, due to the selenium
crate's outdated version of openssl-sys. A quick glance suggests that
it's no longer being maintained, and so I've made the switch here to
`fantoccini` which uses the newer cross-browser WebDriver API. This
crate uses async I/O for communicating with the browser, so this also
includes a dep on tokio for running the futures.

This switch (somehow?) exposed some floating point error introduced
in the test generation. In the course of debugging this issue I moved
the test codegen to use quote/proc_macro2/syn, which ended up fixing
the issue as far as I can tell (none of the off-by-one test errors
happen anymore!).

There's one test which is still marked as ignored, which will be
resolved in a followup commit.