Easy fixes from first round of feedback.
Added `Rect::outset_by` and changed `inset_by(-something)` to use it
instead of `-`.

Made `Layout` and `LayoutOutput` include the overflow fields even when
the content_size feature isn't enabled.  I'm not 100% happy with the
`..default`s I added but it was either that or add lots of
`#[cfg(not(feature="content_size"))]`.  I went with the use of
`..default` as the lesser of two evils.